Hide episode list to avoid whitespace.
The iframe inserts an empty episode list that increases the height
by about 200px, so we end up with a scroll bar unless we set the
height on the player to twice what we need. The added height makes
for a large amount of whitespace at the bottom of the page though,
so it's probably best to just hide the episode windows for now.

When we publish our second episode we'll need to come back and
remove this css so we can see the list again.