Update build-visualization.mdx
Currently when creating a Nerdpack, the default outcome is to add a `nerdlets/` and `launchers/` folder in the Nerdpack. This is unnecessary and potentially confusing if you just want to create a visualization. I've gotten some feedback that it isn't clear you can remove those folders and have the Nerdpack continue working so I'd like to add a note in this guide somewhere. I think where I've added it is a good place but would welcome feedback on that. Thanks!