Clarify steps for publishing a custom viz
When you serve a nerdpack locally for a custom visualization, you can see it right away in the Custom Visualization app (Apps > custom visualizations).  When you officially publish and deploy the custom viz app you have to first subscribe to the nerdpack in the NR1 catalog, log out and log back in, before the custom viz shows up in the custom visualization app.  Myself and other colleagues overlooked this step because when you're testing, you become used to just going straight to the custom visualizations app and having to subscribe to the nerdpack first (after publish and deploy) is no longer intuitive.