Replace config block with link to information
Thanks to @alexronquillo, we now have a [full entry on the configuration options available](https://developer.newrelic.com/explore-docs/custom-viz/configuration-options) and how they map to the UI and visualization user experience. 

Per Alex's suggestion, it makes sense to replace the quick view of configuration with a link to this guide.