chore: removing last year's nerd-days page
This commit removed last year's Nerd Days page(s). We no longer need
them online and there is a bug in the code that is throwing javascript
errors. Additionally, it looks like someone is sending a lot of traffic
to this page right now: https://staging.onenr.io/0znQxp59BQV. This is
causing a lot of on-call pages.

Since there's no value in keeping the page online, I am proposing that
we remove the pages.