chore: added event tracking utility for custom NR Events
NOTE: This is for non-browser "server" events. We don't have the node
agent installed in the Gatsby function "worker" so we need to rely on
New Relic custom events API.

This will require adding our account ID to the environment, in addition
to our API key (which is already in place).