docs: Express & Angular
The Node eng team thought it was important to clarify that this is a full stack app using Express and Angular. Not _just_ an Angular app. Because the Node agent is monitoring the Express BE, and the Browser agent is monitoring Angular FE. The first steps of this are about Express, not Angular. This is significant because there is confusion about when a customer needs the Node agent or just the Browser agent. If this was just an Angular app, there would be no need for both.